password generator
Excited to share that I have completed Task 3 of CodSoft, a Python programming internship. 

🐍 This task involved creating a password generator in Python. 

The program prompts users to input the desired length of the password and generates it accordingly. If the user enters a value below 8, the program notifies them that the password should be a minimum length of 8 characters. 🔒💻

#codsoft #PythonProgramming #Internship #CodingSkills #learningnew